move the declaration of LogTLinInterpScratchBuf to the lnT_prep.hpp header
