+++ src/clib/Make.config.objects
@@ -42,6 +42,7 @@ OBJS_CONFIG_LIB = \
         inject_model/load_data.lo \
         inject_model/raw_data.lo \
         interpolators_g.lo \
+        lnT_prep.lo \
         rate_timestep_g.lo \
         make_consistent.lo \
         scale_fields.lo \
